requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
groq>=0.11.0
plotly>=5.0.0
pandas>=2.0.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
biopython>=1.81
psutil>=5.9.0